pyproject.toml
[project]
name = "autowiki"
version = "0.1.0"
description = "Personal knowledge management markdown files-based wiki with automatic bookkeeping"
readme = "README.md"
authors = [
    { name = "Gilad Barnea", email = "gilad@hear.ai" }
]
requires-python = ">=3.13"
dependencies = [
    "PyYAML>=6.0,<7.0",
]

[project.scripts]
autowiki = "autowiki:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
